==> 01_Ingest_Data.ipynb <==
# Check Pre-Requisites

%store -r setup_instance_check_passed

try:
    setup_instance_check_passed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Instance Check.")
    print("+++++++++++++++++++++++++++++++")

print(setup_instance_check_passed)

%store -r setup_dependencies_passed

try:
    setup_dependencies_passed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup Dependencies.")
    print("+++++++++++++++++++++++++++++++")

print(setup_dependencies_passed)

%store -r setup_s3_bucket_passed

try:
    setup_s3_bucket_passed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup S3 Bucket.")
    print("+++++++++++++++++++++++++++++++")

print(setup_s3_bucket_passed)

%store -r setup_iam_roles_passed

try:
    setup_iam_roles_passed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup IAM Roles.")
    print("+++++++++++++++++++++++++++++++")

print(setup_iam_roles_passed)

if not setup_instance_check_passed:
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Instance Check.")
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
if not setup_dependencies_passed:
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup Dependencies.")
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
if not setup_s3_bucket_passed:
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup S3 Bucket.")
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
if not setup_iam_roles_passed:
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL NOTEBOOKS IN THE SETUP FOLDER FIRST. You are missing Setup IAM Roles.")
    print("+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++")

import boto3
import sagemaker
import pandas as pd

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name
account_id = boto3.client("sts").get_caller_identity().get("Account")

sm = boto3.Session().client(service_name="sagemaker", region_name=region)

---

# Create Bucket

!aws s3 mb s3://ads-508-team4

# Set S3 Source Location (Public S3 Bucket)

s3_public_path_csv = "s3://ads-508-team4/olist/csv"

%store s3_public_path_csv

# Set S3 Destination Location (Our Private S3 Bucket)

s3_private_path_csv = "s3://ads-508-team4/olist/csv".format(bucket)
print(s3_private_path_csv)

%store s3_private_path_csv

# Create folders for each datasets
#### Total 4 datasets for the project

1. olist_customers_dataset.csv

s3_cust_path_csv = "s3://ads-508-team4/olist/csv/cust".format(bucket)
print(s3_cust_path_csv)

2. olist_order_items_dataset.csv

s3_order_items_path_csv = "s3://ads-508-team4/olist/csv/order_items".format(bucket)
print(s3_order_items_path_csv)

3. olist_orders_dataset.csv

s3_order_path_csv = "s3://ads-508-team4/olist/csv/order".format(bucket)
print(s3_order_path_csv)

4. olist_products_dataset.csv

s3_products_path_csv = "s3://ads-508-team4/olist/csv/products".format(bucket)
print(s3_products_path_csv)

## --- Dataset Manually Uploaded ---

# List Folders in our Private S3 Bucket in this Account

!aws s3 ls $s3_private_path_csv/

from IPython.core.display import display, HTML

display(
    HTML(
        '<b>Review <a target="blank" href="https://s3.console.aws.amazon.com/s3/buckets/ads-508-team4/olist/csv/?region={}&tab=overview">S3 Bucket</a></b>'.format(
            region, account_id, region
        )
    )
)

# Store Variables for the Next Notebooks

%store

---

# Create Athena Database Schema

import boto3
import sagemaker

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

ingest_create_athena_db_passed = False

%store -r s3_public_path_csv

try:
    s3_public_path_csv
except NameError:
    print("*****************************************************************************")
    print("[ERROR] PLEASE RE-RUN THE PREVIOUS COPY TSV TO S3 NOTEBOOK ******************")
    print("[ERROR] THIS NOTEBOOK WILL NOT RUN PROPERLY. ********************************")
    print("*****************************************************************************")

print(s3_public_path_csv)

%store -r s3_private_path_csv

try:
    s3_private_path_csv
except NameError:
    print("*****************************************************************************")
    print("[ERROR] PLEASE RE-RUN THE PREVIOUS COPY TSV TO S3 NOTEBOOK ******************")
    print("[ERROR] THIS NOTEBOOK WILL NOT RUN PROPERLY. ********************************")
    print("*****************************************************************************")

print(s3_private_path_csv)

# Import PyAthena

# !pip install --disable-pip-version-check -q PyAthena==2.1.0

from pyathena import connect

database_name = "olist"

# Set S3 staging directory -- this is a temporary directory used for Athena queries
s3_staging_dir = "s3://ads-508-team4/athena/staging".format(bucket)

conn = connect(region_name=region, s3_staging_dir=s3_staging_dir)

statement = "CREATE DATABASE IF NOT EXISTS {}".format(database_name)
print(statement)

import pandas as pd

pd.read_sql(statement, conn)

# Verify The Database Has Been Created Succesfully

statement = "SHOW DATABASES"

df_show = pd.read_sql(statement, conn)
df_show.head(5)

if database_name in df_show.values:
    ingest_create_athena_db_passed = True

%store ingest_create_athena_db_passed

# Store Variables for the Next Notebooks

%store

---

# Register TSV Data With Athena

import boto3
import sagemaker
import pandas as pd

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

ingest_create_athena_table_csv_passed = False

%store -r ingest_create_athena_db_passed

try:
    ingest_create_athena_db_passed
except NameError:
    print("++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL PREVIOUS NOTEBOOKS.  You did not create the Athena Database.")
    print("++++++++++++++++++++++++++++++++++++++++++++++")

print(ingest_create_athena_db_passed)

if not ingest_create_athena_db_passed:
    print("++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN ALL PREVIOUS NOTEBOOKS.  You did not create the Athena Database.")
    print("++++++++++++++++++++++++++++++++++++++++++++++")
else:
    print("[OK]")

%store -r s3_private_path_csv

print(s3_private_path_csv)

# Create Athena Table from Local CSV Files

# Set S3 staging directory -- this is a temporary directory used for Athena queries
s3_staging_dir = "s3://ads-508-team4/athena/staging".format(bucket)

# Set Athena parameters
    # database name
database_name = "olist"
    # table names
cust_table_name = "olist_customers_dataset"
order_items_table_name = "olist_order_items_dataset"
order_table_name = "olist_order_dataset"
products_table_name = "olist_products_dataset"

conn = connect(region_name=region, s3_staging_dir=s3_staging_dir)

print(s3_cust_path_csv)
print(s3_order_items_path_csv)
print(s3_order_path_csv)
print(s3_products_path_csv)

## 1. Table for olist_customers_dataset

create_cust_tb = """

CREATE EXTERNAL TABLE IF NOT EXISTS  {}.{}
(
         customer_id                  string,
         customer_unique_id           string,
         customer_zip_code_prefix     string,
         customer_city                string,
         customer_state               string
)  

ROW FORMAT DELIMITED FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\\n' 
STORED AS TEXTFILE
LOCATION  '{}'
TBLPROPERTIES ('compressionType'='gzip', 'skip.header.line.count'='1')

""".format(database_name, cust_table_name, s3_cust_path_csv)

pd.read_sql(create_cust_tb, conn)

### Verify The Table Has Been Created Succesfully 

pd.read_sql("""SELECT * FROM olist.olist_customers_dataset LIMIT 10 """, conn)

pd.read_sql("""select count(customer_id) from olist.olist_customers_dataset """, conn)

## 2. Table for order_items_table_name

create_order_items_tb = """
CREATE EXTERNAL TABLE IF NOT EXISTS  {}.{}
(
         order_id                string,
         order_item_id           string,
         product_id              string,
         seller_id               string,
         shipping_limit_date     timestamp,
         price                   decimal(10,2),
         freight_value           decimal(10,2)
)  

ROW FORMAT DELIMITED FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\\n' 
STORED AS TEXTFILE
LOCATION  '{}'
TBLPROPERTIES ('compressionType'='gzip', 'skip.header.line.count'='1')

""".format(database_name, order_items_table_name, s3_order_items_path_csv)

pd.read_sql(create_order_items_tb, conn)

### Verify The Table Has Been Created Succesfully 

pd.read_sql("""SELECT * FROM olist.olist_order_items_dataset LIMIT 10 """, conn)

pd.read_sql("""select count(*) from olist.olist_order_items_dataset """, conn)

## 3. Table for olist_orders_dataset

create_order_tb = """
CREATE EXTERNAL TABLE IF NOT EXISTS  {}.{}
(
         order_id                       string,
         customer_id                    string,
         order_status                   string,
         order_purchase_timestamp       timestamp,
         order_approved_at              timestamp,
         order_delivered_carrier_date   timestamp,
         order_delivered_customer_date  timestamp,
         order_estimated_delivery_date  timestamp
)  

ROW FORMAT DELIMITED FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\\n' 
STORED AS TEXTFILE
LOCATION  '{}'
TBLPROPERTIES ('compressionType'='gzip', 'skip.header.line.count'='1')

""".format(database_name, order_table_name, s3_order_path_csv)

pd.read_sql(create_order_tb, conn)

### Verify The Table Has Been Created Succesfully 

pd.read_sql("""SELECT * FROM olist.olist_order_dataset LIMIT 10 """, conn)

pd.read_sql("""select count(*) from olist.olist_order_dataset """, conn)

## 4. Table for olist_products_dataset

create_products_tb = """
CREATE EXTERNAL TABLE IF NOT EXISTS  {}.{}
(
         product_id                   string,
         product_category_name        string,
         product_name_lenght          int,
         product_description_lenght   int,
         product_photos_qty           int,
         product_weight_g             int,
         product_length_cm            int,
         product_height_cm            int,
         product_width_cm             int
)  

ROW FORMAT DELIMITED FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\\n' 
STORED AS TEXTFILE
LOCATION  '{}'
TBLPROPERTIES ('compressionType'='gzip', 'skip.header.line.count'='1')

""".format(database_name, products_table_name, s3_products_path_csv)

pd.read_sql(create_products_tb, conn)

### Verify The Table Has Been Created Succesfully 

pd.read_sql("""SELECT * FROM olist.olist_products_dataset LIMIT 10 """, conn)

pd.read_sql("""select count(*) from olist.olist_products_dataset """, conn)

# Review the New Athena Table in the Glue Catalog

from IPython.core.display import display, HTML

display(
    HTML(
        '<b>Review <a target="top" href="https://console.aws.amazon.com/glue/home?region={}#">AWS Glue Catalog</a></b>'.format(
            region
        )
    )
)

# Store Variables for the Next Notebooks

%store

%%html

<p><b>Shutting down your kernel for this notebook to release resources.</b></p>
<button class="sm-command-button" data-commandlinker-command="kernelmenu:shutdown" style="display:none;">Shutdown Kernel</button>
        
<script>
try {
    els = document.getElementsByClassName("sm-command-button");
    els[0].click();
}
catch(err) {
    // NoOp
}    
</script>

%%javascript

try {
    Jupyter.notebook.save_checkpoint();
    Jupyter.notebook.session.delete();
}
catch(err) {
    // NoOp
}
