==> 03_Prepare_Bias.ipynb <==
# Prepare Dataset for Bias Analysis

import boto3
import sagemaker
import pandas as pd

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

!aws s3 cp 's3://ads-508-team4/olist/csv/cust/olist_customers_dataset.csv' ./data-clarify/
!aws s3 cp 's3://ads-508-team4/olist/csv/order_items/olist_order_items_dataset.csv' ./data-clarify/
!aws s3 cp 's3://ads-508-team4/olist/csv/order/olist_orders_dataset.csv' ./data-clarify/
!aws s3 cp 's3://ads-508-team4/olist/csv/products/olist_products_dataset.csv' ./data-clarify/

import csv

df_cust = pd.read_csv(
    "./data-clarify/olist_customers_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)
df_cust.shape
df_cust.head()

order_items = pd.read_csv(
    "./data-clarify/olist_order_items_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)
order_items.shape
order_items.head()

order = pd.read_csv(
    "./data-clarify/olist_orders_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)
order.shape
order.head()

prod = pd.read_csv(
    "./data-clarify/olist_products_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)
prod.shape
prod.head()

# Combine Dataframe

df1 = df_cust.merge(order, how='outer', on='"customer_id"')

df2 = df1.merge(order_items, how='outer', on='"order_id"')

df3 = df2.merge(prod, how='outer', on='"product_id"' )

df3.head()

# Balance the Dataset by Product Category and Price

df_grouped_by = df3.groupby(['"product_category_name"', '"price"'])[['"product_category_name"', '"price"']]
df_balanced = df_grouped_by.apply(lambda x: x.sample(df_grouped_by.size().min()).reset_index(drop=True))
df_balanced.shape

### Unbalanced label classes

path = "./data-clarify/master_unbal.csv"
df3.to_csv(path, index=False, header=True)

### Balanced label classes

path_balanced = "./data-clarify/master_bal.csv"
df_balanced.to_csv(path_balanced, index=False, header=True)

### Write as JSONLINES

path_jsonlines = "./data-clarify/master_bal.jsonl"
df_balanced.to_json(path_or_buf=path_jsonlines, orient="records", lines=True)

# Upload Train Data to S3¶

import time

timestamp = int(time.time())

bias_data_s3_uri = sess.upload_data(bucket="ads-508-team4", key_prefix="bias-detection-{}".format(timestamp), path=path)
bias_data_s3_uri

!aws s3 ls $bias_data_s3_uri

balanced_bias_data_s3_uri = sess.upload_data(
    bucket="ads-508-team4", key_prefix="bias-detection-{}".format(timestamp), path=path_balanced
)
balanced_bias_data_s3_uri

!aws s3 ls $balanced_bias_data_s3_uri

balanced_bias_data_jsonlines_s3_uri = sess.upload_data(
    bucket="ads-508-team4", key_prefix="bias-detection-{}".format(timestamp), path=path_jsonlines
)
balanced_bias_data_jsonlines_s3_uri

!aws s3 ls $balanced_bias_data_jsonlines_s3_uri

# Store Variables for Next Notebook(s)

%store balanced_bias_data_jsonlines_s3_uri

%store balanced_bias_data_s3_uri

%store bias_data_s3_uri

%store
