==> 04_Run_Bias_Analysis.ipynb <==
# Run Ad-Hoc Data Bias Analysis

from smclarify.bias import report
from typing import Dict
from collections import defaultdict
import pandas as pd
import seaborn as sns

# Read Dataset From S3

%store -r bias_data_s3_uri

print(bias_data_s3_uri)

%store -r balanced_bias_data_s3_uri

print(balanced_bias_data_s3_uri)

!aws s3 cp $bias_data_s3_uri ./data-clarify/

!aws s3 cp $balanced_bias_data_s3_uri ./data-clarify/

df = pd.read_csv("./data-clarify/master_bal.csv")
df.shape

# Calculate Bias Metrics on Unbalanced Data

facet_column = report.FacetColumn(name='"product_category_name"')

label_column = report.LabelColumn(
    name='"price"', 
    data=df['"price"'],
    positive_label_values=[150, ]
)

# Run SageMaker Clarify Bias Report

report.bias_report(
    df=df, 
    facet_column=facet_column, 
    label_column=label_column, 
    stage_type=report.StageType.PRE_TRAINING, 
    metrics=["CI", "DPL", "KL", "JS", "LP", "TVD", "KS"]
)

%%html

<p><b>Shutting down your kernel for this notebook to release resources.</b></p>
<button class="sm-command-button" data-commandlinker-command="kernelmenu:shutdown" style="display:none;">Shutdown Kernel</button>
        
<script>
try {
    els = document.getElementsByClassName("sm-command-button");
    els[0].click();
}
catch(err) {
    // NoOp
}    
</script>
