==> 05_Data_Quailty.ipynb <==
%store -r ingest_create_athena_table_csv_passed

try:
    ingest_create_athena_table_csv_passed
except NameError:
    print("++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN THE NOTEBOOKS IN THE INGEST FOLDER FIRST. You did not register the TSV Data.")
    print("++++++++++++++++++++++++++++++++++++++++++++++")

print(ingest_create_athena_table_csv_passed)

if not ingest_create_athena_table_csv_passed:
    print("++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] YOU HAVE TO RUN THE NOTEBOOKS IN THE INGEST FOLDER FIRST. You did not register the TSV Data.")
    print("++++++++++++++++++++++++++++++++++++++++++++++")
else:
    print("[OK]")

!pygmentize preprocess-deequ-pyspark.py

import sagemaker
import boto3

sess = sagemaker.Session()
# bucket = sess.default_bucket()
bucket = 'ads-508-team4'
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

from sagemaker.spark.processing import PySparkProcessor

processor = PySparkProcessor(
    base_job_name="olist-analyzer",
    role=role,
    framework_version="2.4",
    instance_count=1,
    instance_type="ml.r5.2xlarge",
    max_runtime_in_seconds=300,
)

s3_input_data = "s3://ads-508-team4/olist/csv/"
print(s3_input_data)

!aws s3 ls $s3_input_data

# Setup Output Data

from time import gmtime, strftime

timestamp_prefix = strftime("%Y-%m-%d-%H-%M-%S", gmtime())

output_prefix = "olist-analyzer-{}".format(timestamp_prefix)
processing_job_name = "olist-analyzer-{}".format(timestamp_prefix)

print("Processing job name:  {}".format(processing_job_name))

s3_output_analyze_data = "s3://ads-508-team4/{}/output".format(output_prefix)

print(s3_output_analyze_data)
