==> 06_Prepare_Dataset.ipynb <==
import sagemaker
import boto3
import csv
import numpy as np
import pandas as pd

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

sm = boto3.Session().client(service_name="sagemaker", region_name=region)
s3 = boto3.Session().client(service_name="s3", region_name=region)

pd.set_option('display.max_columns', None)

# Combine Data

!aws s3 cp 's3://ads508-team4/olist/csv/cust/olist_customers_dataset.csv' ./data-clarify/
!aws s3 cp 's3://ads508-team4/olist/csv/order_items/olist_order_items_dataset.csv' ./data-clarify/
!aws s3 cp 's3://ads508-team4/olist/csv/order/olist_orders_dataset.csv' ./data-clarify/
!aws s3 cp 's3://ads508-team4/olist/csv/products/olist_products_dataset.csv' ./data-clarify/

### 1. Customer dataset

df_cust = pd.read_csv(
    "./data-clarify/olist_customers_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)

df_cust.columns=df_cust.columns.str.replace('"','')
df_cust['customer_id'] = df_cust['customer_id'].str.replace('"','')
df_cust['customer_unique_id'] = df_cust['customer_unique_id'].str.replace('"','')
df_cust['customer_zip_code_prefix'] = df_cust['customer_zip_code_prefix'].str.replace('"','')

df_cust.head()

### 2. Order items dataset

order_items = pd.read_csv(
    "./data-clarify/olist_order_items_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)

order_items.columns=order_items.columns.str.replace('"','')
order_items['order_id'] = order_items['order_id'].str.replace('"','')
order_items['product_id'] = order_items['product_id'].str.replace('"','')
order_items['seller_id'] = order_items['seller_id'].str.replace('"','')

order_items.head()

### 3. Order dataset

order = pd.read_csv(
    "./data-clarify/olist_orders_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)

order = pd.read_csv(
    "./data-clarify/olist_orders_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)
order.columns=order.columns.str.replace('"','')
order['order_id'] = order['order_id'].str.replace('"','')
order['customer_id'] = order['customer_id'].str.replace('"','')

order.head()

### 4. Product dataset

prod = pd.read_csv(
    "./data-clarify/olist_products_dataset.csv",
    delimiter=",",
    quoting=csv.QUOTE_NONE
)

prod.columns=prod.columns.str.replace('"','')
prod['product_id'] = prod['product_id'].str.replace('"','')

prod.head()

# Merge datasets

order_cust = order.merge(df_cust, how='left', on='customer_id' )
order_cust_item = order_cust.merge(order_items, how='left', on='order_id')
master = order_cust_item.merge(prod, how='left', on='product_id')

# Transform product categoreis and translate to English

master['agro_industry_and_commerce'] = np.where(master['product_category_name'].isin(['agro_industria_e_comercio']), 1, 0)
master['foods'] = np.where(master['product_category_name'].isin(['alimentos','alimentos_bebidas','la_cuisine']), 1 ,0)
master['arts'] = np.where(master['product_category_name'].isin(['artes','artes_e_artesanato']),1 ,0)
master['party'] = np.where(master['product_category_name'].isin(['artigos_de_festas','artigos_de_natal']), 1 ,0)
master['audio'] = np.where(master['product_category_name'].isin(['audio']), 1 ,0)
master['automotive'] = np.where(master['product_category_name'].isin(['automotivo']), 1 ,0)
master['babies'] = np.where(master['product_category_name'].isin(['bebes','fraldas_higiene']), 1 ,0)
master['drinks'] = np.where(master['product_category_name'].isin(['bebidas']), 1 ,0)
master['beauty_health'] = np.where(master['product_category_name'].isin(['beleza_saude']), 1 ,0)
master['toys'] = np.where(master['product_category_name'].isin(['brinquedos']), 1 ,0)

master['bed_table_bath'] = np.where(master['product_category_name'].isin(['cama_mesa_banho']), 1 ,0)
master['home'] = np.where(master['product_category_name'].isin(['casa_conforto','casa_conforto_2','casa_construcao', 'utilidades_domesticas']), 1 ,0)
master['music_cds_dvds'] = np.where(master['product_category_name'].isin(['cds_dvds_musicais']), 1 ,0)
master['cine_photo'] = np.where(master['product_category_name'].isin(['cine_foto']), 1 ,0)
master['air_conditioning'] = np.where(master['product_category_name'].isin(['climatizacao']), 1 ,0)
master['game_consoles'] = np.where(master['product_category_name'].isin(['consoles_games']), 1 ,0)
master['construction'] = np.where(master['product_category_name'].isin(['construcao_ferramentas_construcao','construcao_ferramentas_ferramentas','construcao_ferramentas_iluminacao','construcao_ferramentas_jardim','construcao_ferramentas_seguranca']), 1 ,0)
master['cool_stuff'] = np.where(master['product_category_name'].isin(['cool_stuff','seguros_e_servicos','sinalizacao_e_seguranca']), 1 ,0)
master['blu_ray_dvds'] = np.where(master['product_category_name'].isin(['dvds_blu_ray']), 1 ,0)
master['electronics'] = np.where(master['product_category_name'].isin(['eletrodomesticos','eletrodomesticos_2', 'eletronicos', 'eletroportateis']), 1 ,0)

master['sport_leisure'] = np.where(master['product_category_name'].isin(['esporte_lazer']), 1 ,0)
master['fashion'] = np.where(master['product_category_name'].isin(['fashion_bolsas_e_acessorios','fashion_calcados','fashion_esporte','fashion_roupa_feminina','fashion_roupa_infanto_juvenil','fashion_roupa_masculina','fashion_underwear_e_moda_praia']), 1 ,0)
master['garden'] = np.where(master['product_category_name'].isin(['ferramentas_jardim','flores']), 1 ,0)
master['industry'] = np.where(master['product_category_name'].isin(['industria_comercio_e_negocios']), 1 ,0)
master['computer_accessories'] = np.where(master['product_category_name'].isin(['informatica_acessorios']), 1 ,0)
master['musical_instruments'] = np.where(master['product_category_name'].isin(['instrumentos_musicais']), 1 ,0)
master['books'] = np.where(master['product_category_name'].isin(['livros_importados','livros_interesse_geral','livros_tecnicos']), 1 ,0)
master['accessory_bags'] = np.where(master['product_category_name'].isin(['malas_acessorios']), 1 ,0)
master['market_place'] = np.where(master['product_category_name'].isin(['market_place']), 1 ,0)
master['furniture'] = np.where(master['product_category_name'].isin(['moveis_colchao_e_estofado','moveis_cozinha_area_de_servico_jantar_e_jardim','moveis_decoracao','moveis_escritorio','moveis_quarto','moveis_sala']), 1 ,0)

master['song'] = np.where(master['product_category_name'].isin(['musica']), 1 ,0)
master['stationery'] = np.where(master['product_category_name'].isin(['papelaria']), 1 ,0)
master['pcs'] = np.where(master['product_category_name'].isin(['pc_gamer','pcs', 'portateis_casa_forno_e_cafe','portateis_cozinha_e_preparadores_de_alimentos']), 1 ,0)
master['perfume'] = np.where(master['product_category_name'].isin(['perfumaria']), 1 ,0)
master['pet'] = np.where(master['product_category_name'].isin(['pet_shop']), 1 ,0)
master['gifts_watches'] = np.where(master['product_category_name'].isin(['relogios_presentes']), 1 ,0)
master['insurance_and_services'] = np.where(master['product_category_name'].isin(['seguros_e_servicos']), 1 ,0)
master['phone'] = np.where(master['product_category_name'].isin(['telefonia','telefonia_fixa']), 1 ,0)

master.head()

# Transform dataset to customer level

cust_master = master.groupby(['customer_unique_id']).agg({
'order_id'                       : 'count',
'price'                          : 'sum',
'freight_value'                  : 'sum',
'product_weight_g'               : 'sum',
'agro_industry_and_commerce'     : 'sum',
'foods'                          : 'sum',
'arts'                           : 'sum',
'party'                          : 'sum',
#'articles'                       : 'sum',
'audio'                          : 'sum', 
'automotive'                     : 'sum', 
'babies'                         : 'sum', 
'drinks'                         : 'sum',  
'beauty_health'                  : 'sum',
'toys'                           : 'sum',
#'bed table bath'                 : 'sum',
'bed_table_bath'                 : 'sum',
'home'                           : 'sum', 
'music_cds_dvds'                 : 'sum',
'cine_photo'                     : 'sum',
'air_conditioning'               : 'sum',
'game_consoles'                  : 'sum', 
'construction'                   : 'sum',
'cool_stuff'                     : 'sum',
'blu_ray_dvds'                   : 'sum',
#'home_appliances'                : 'sum',
'electronics'                    : 'sum',
'sport_leisure'                  : 'sum',
'fashion'                        : 'sum', 
'garden'                         : 'sum',
'industry'                       : 'sum',
'computer_accessories'           : 'sum',
'musical_instruments'            : 'sum', 
'books'                          : 'sum',
'accessory_bags'                 : 'sum',
'market_place'                   : 'sum', 
'furniture'                      : 'sum', 
'song'                           : 'sum', 
'stationery'                     : 'sum',
'pcs'                            : 'sum', 
'perfume'                        : 'sum',
'pet'                            : 'sum', 
'gifts_watches'                  : 'sum',
'insurance_and_services'         : 'sum',
'phone'                          : 'sum'
}).reset_index()

print(cust_master.shape)
cust_master.head()

cust_master.columns = ['customer_unique_id', 'order_cnt',
                       'total_spend', 'total_shipping', 'total_prod_weight_g',
                       'agro_industry_and_commerce', 'foods', 'arts', 'party', 'audio',
                       'automotive', 'babies', 'drinks', 'beauty_health', 'toys',
                       'bed_table_bath', 'home', 'music_cds_dvds',
                       'cine_photo', 'air_conditioning', 'game_consoles', 'construction',
                       'cool_stuff', 'blu_ray_dvds', 'electronics',
                       'sport_leisure', 'fashion', 'garden',  'industry',
                       'computer_accessories', 'musical_instruments',  'books',
                       'accessory_bags', 'market_place', 'furniture', 'song', 'stationery',
                       'pcs', 'perfume', 'pet',  'gifts_watches',
                       'insurance_and_services', 'phone']

## Latest state and city infomation

demo = master.sort_values(['order_purchase_timestamp'], ascending=False).groupby(['customer_unique_id']).head(1)
# demo = demo[['customer_unique_id', 'customer_city', 'customer_state' ]]
demo = demo[['customer_unique_id', 'customer_state' ]]

print(demo.shape)
demo.head()

## Merge state to the master dataset

cust_master = cust_master.merge(demo, how='left', on='customer_unique_id' )

cust_master['SP'] = np.where(cust_master['customer_state'] == 'SP', 1, 0)
cust_master['SC'] = np.where(cust_master['customer_state'] == 'SC', 1, 0)
cust_master['PA'] = np.where(cust_master['customer_state'] == 'PA', 1, 0)
cust_master['PR'] = np.where(cust_master['customer_state'] == 'PR', 1, 0)
cust_master['RJ'] = np.where(cust_master['customer_state'] == 'RJ', 1, 0)

cust_master['MA'] = np.where(cust_master['customer_state'] == 'MA', 1, 0)
cust_master['ES'] = np.where(cust_master['customer_state'] == 'ES', 1, 0)
cust_master['RS'] = np.where(cust_master['customer_state'] == 'RS', 1, 0)
cust_master['MG'] = np.where(cust_master['customer_state'] == 'MG', 1, 0)
cust_master['GO'] = np.where(cust_master['customer_state'] == 'GO', 1, 0)

cust_master['DF'] = np.where(cust_master['customer_state'] == 'DF', 1, 0)
cust_master['BA'] = np.where(cust_master['customer_state'] == 'BA', 1, 0)
cust_master['MT'] = np.where(cust_master['customer_state'] == 'MT', 1, 0)
cust_master['SE'] = np.where(cust_master['customer_state'] == 'SE', 1, 0)
cust_master['TO'] = np.where(cust_master['customer_state'] == 'TO', 1, 0)

cust_master['PE'] = np.where(cust_master['customer_state'] == 'PE', 1, 0)
cust_master['CE'] = np.where(cust_master['customer_state'] == 'CE', 1, 0)
cust_master['AL'] = np.where(cust_master['customer_state'] == 'AL', 1, 0)
cust_master['AC'] = np.where(cust_master['customer_state'] == 'AC', 1, 0)
cust_master['PI'] = np.where(cust_master['customer_state'] == 'PI', 1, 0)

cust_master['PB'] = np.where(cust_master['customer_state'] == 'PB', 1, 0)
cust_master['RN'] = np.where(cust_master['customer_state'] == 'RN', 1, 0)
cust_master['AM'] = np.where(cust_master['customer_state'] == 'AM', 1, 0)
cust_master['MS'] = np.where(cust_master['customer_state'] == 'MS', 1, 0)
cust_master['RR'] = np.where(cust_master['customer_state'] == 'RR', 1, 0)

cust_master['AP'] = np.where(cust_master['customer_state'] == 'AP', 1, 0)
cust_master['RO'] = np.where(cust_master['customer_state'] == 'RO', 1, 0)

cust_master = cust_master.drop(['customer_state','customer_unique_id'], axis=1)

cust_master.head()

# Remove outliers

from scipy import stats
cust_remove_out = cust_master[(np.abs(stats.zscore(cust_master['total_spend'])) < 3)]
cust_remove_out.shape

cust_remove_out['total_spend'].describe()

# Set target variable

cust_master['spending_grp'] = pd.cut(cust_master['total_spend'], 
                                    [0, 45.9, 86.895, 149.9, np.inf],
                                   # labels=["Normal", "Average", "Good", "Excellent"])
                                    labels=["1", "2", "3", "4"])

cust_master['spending_grp'].value_counts()

cols = list(cust_master)
cols.insert(0, cols.pop(cols.index('spending_grp')))
cust_master = cust_master.loc[:, cols]
cust_master.head()

target = cust_master['spending_grp']
%store target

# Split datasets into train, test, and validate

* 60% train
* 20% validate
* 20% test

from sklearn.model_selection import train_test_split

train_data, validate_data, test_data = np.split(cust_master.sample(frac=1, random_state=42), 
                        [int(.6*len(cust_master)), int(.8*len(cust_master))])

print(train_data.shape)
print(validate_data.shape)
print(test_data.shape)

s3_private_path_data = "s3://ads508-team4/modeling_groups".format(bucket)
print(s3_private_path_data)

%store s3_private_path_data

train_file = "train_data.csv"
train_data.to_csv(train_file, index=False, header=True)

validate_file = "validate_data.csv"
validate_data.to_csv(validate_file, index=False, header=True)

test_file = "test_data.csv"
test_data.to_csv(test_file, index=False, header=True)

# Getting files to s3 bucket
train_data_s3_path = sess.upload_data(bucket="ads508-team4", key_prefix="modeling_groups", path=train_file)
validate_data_s3_path = sess.upload_data(bucket="ads508-team4", key_prefix="modeling_groups", path=validate_file)
test_data_s3_path = sess.upload_data(bucket="ads508-team4", key_prefix="modeling_groups", path=test_file)

!aws s3 ls $modeling_groups

%store

%%html

<p><b>Shutting down your kernel for this notebook to release resources.</b></p>
<button class="sm-command-button" data-commandlinker-command="kernelmenu:shutdown" style="display:none;">Shutdown Kernel</button>
        
<script>
try {
    els = document.getElementsByClassName("sm-command-button");
    els[0].click();
}
catch(err) {
    // NoOp
}    
</script>

Shutting down your kernel for this notebook to release resources.

%%javascript

try {
    Jupyter.notebook.save_checkpoint();
    Jupyter.notebook.session.delete();
}
catch(err) {
    // NoOp
}
